用户列表查询 新增用户时用到的方法

+++ blade-system/src/main/java/com/blade/manager/system/permission/mapper/MenuMapper.java
@@ -24,7 +24,7 @@ public interface MenuMapper extends BaseMapper<Menu> {
      * @param roleIds 角色ID
      * @return menus
      */
-    List<Menu> selectMenuByRoleIds(@Param("roleIds") List<Integer> roleIds);
+    List<Menu> selectMenuByRoleIds(@Param("roleIds") List<Long> roleIds);
 
     /**
      * 获取全部菜单
